refactor(oc:8284): estrai trackLiveDistanceVm in un selettore condiviso
La logica di combinazione trackDistanceCovered/trackRemainingDistance/
trackPositionStale + gate su confOPTIONSShowTrackRemainingDistance era
duplicata identica tra tab-detail.component.ts (oc:8177) e
track-recorder.component.ts (oc:8284, repo principale). Estratta in un
unico selettore in user-activity.selector.ts, riusato da entrambi.

Trovato in review formale (wm-review-ticket).
